+++ packages/common/src/Components/ToolbarButton.jsx
@@ -8,22 +8,17 @@ class ToolbarButton extends Component {
     showTooltip: PropTypes.bool,
     tooltipText: PropTypes.string,
     button: PropTypes.element,
+    tooltipOffset: PropTypes.shape({
+      x: PropTypes.number,
+      y: PropTypes.number
+    })
   };
 
 
   render() {
-    const { theme, showTooltip, tooltipText, button } = this.props;
-
+    const { theme, showTooltip, tooltipText, button, tooltipOffset } = this.props;
     if (showTooltip) {
-      return (
-        <Tooltip
-          content={tooltipText}
-          moveBy={{ x: 12, y: 0 }}
-          theme={theme}
-        >
-          {button}
-        </Tooltip>
-      );
+      return <Tooltip content={tooltipText} moveBy={tooltipOffset} theme={theme}>{button}</Tooltip>;
     } else {
       return button;
     }