+++ packages/editor/src/RichContentEditor/Toolbars/buttons/TextButton.jsx
@@ -67,6 +67,6 @@ export default class TextButton extends Component {
     );
     /* eslint-enable jsx-a11y/no-static-element-interactions */
 
-    return <ToolbarButton theme={theme} showTooltip={showTooltip} tooltipText={tooltipText} button={textButton} />;
+    return <ToolbarButton theme={theme} showTooltip={showTooltip} tooltipText={tooltipText} button={textButton} tooltipOffset={{ y: -20 }} />;
   }
 }